+++ @planetarium/account-web3-secret-storage/src/Web3KeyStore.ts
@@ -3,6 +3,7 @@ import { PassphraseEntry } from "./PassphraseEntry.js";
 import {
   encryptKeyObject,
   isKeyObject,
+  type DecryptionOptions,
   Web3Account,
   Web3KeyObject,
 } from "./Web3Account.js";