+++ @planetarium/account-web3-secret-storage/test/Web3Account.test.ts
@@ -1,15 +1,18 @@
-import * as fc from "fast-check";
 import { PublicKey, RawPrivateKey } from "@planetarium/account";
+import * as fc from "fast-check";
+import * as crypto from "node:crypto";
+import { describe, expect, test } from "vitest";
 import { KeyId } from "../src/KeyId";
 import { Passphrase } from "../src/PassphraseEntry";
 import {
+  DecryptionOptions,
+  decryptKeyObject,
+  encryptKeyObject,
+  IncorrectPassphraseError,
+  WeakPrivateKeyError,
   Web3Account,
   Web3KeyObject,
-  encryptKeyObject,
-  decryptKeyObject,
-  IncorrectPassphraseError,
 } from "../src/Web3Account";
-import { describe, expect, test } from "vitest";
 import { MockPassphraseEntry } from "./MockPassphraseEntry";
 
 const pbkdf2KeyObject: Web3KeyObject = {
@@ -57,7 +60,35 @@ const scryptKeyObject: Web3KeyObject = {
   version: 3,
 };
 
-const testVectors = {
+const insufficientLengthKeyObject: Web3KeyObject = {
+  crypto: {
+    cipher: "aes-128-ctr",
+    cipherparams: {
+      iv: "24ece9686a10f8687dc70ca42873a063",
+    },
+    ciphertext: "14",
+    kdf: "pbkdf2",
+    kdfparams: {
+      c: 10240,
+      dklen: 32,
+      prf: "hmac-sha256",
+      salt: "99b480642f6406a10c67429168fb36ba35da7743965e10390eda649ec6da2ae8",
+    },
+    mac: "7856d88908c72f0af78457ccd085696ad6193f7d2bda525723862f450ce060fa",
+  },
+  id: "b35a2647-8581-43ff-a98e-6083dc952632",
+  address: "eb9afe072c781401bf364224c75a036e4d832f52",
+  version: 3,
+};
+
+interface TestVector {
+  keyObject: Web3KeyObject;
+  passphrase: Passphrase;
+  privateKey: RawPrivateKey;
+  options?: Partial<DecryptionOptions>;
+}
+
+const testVectors: Record<string, TestVector> = {
   pbkdf: {
     keyObject: pbkdf2KeyObject,
     passphrase: "passphrase",
@@ -72,53 +103,146 @@ const testVectors = {
       "7a28b5ba57c53603b0b07b56bba752f7784bf506fa95edc395f5cf6c7514fe9d",
     ),
   },
+  insufficientLength: {
+    keyObject: insufficientLengthKeyObject,
+    passphrase: "1",
+    privateKey: RawPrivateKey.fromHex(
+      "00000000000000000000000000000000000000000000000000000000000000aa",
+    ),
+    options: {
+      allowWeakPrivateKey: true,
+    },
+  },
 };
 
 describe("Web3Account", () => {
-  test("getPublicKey()", async () => {
-    const passphraseEntry = new MockPassphraseEntry();
-    const key = new Web3Account(pbkdf2KeyObject, passphraseEntry);
+  describe("getPublicKey()", () => {
+    test("fastcheck", async () => {
+      const passphraseEntry = new MockPassphraseEntry();
+      const key = new Web3Account(pbkdf2KeyObject, passphraseEntry);
 
-    expect(passphraseEntry.authenticateCalls).toStrictEqual([]);
-    expect(passphraseEntry.configurePassphraseCalls).toBe(0);
-    const publicKey = await key.getPublicKey();
-    expect(passphraseEntry.authenticateCalls).toStrictEqual([
-      { keyId: "3c7bec5e-1f1d-4754-a1ce-3644ce1130f1", firstAttempt: true },
-    ]);
-    expect(passphraseEntry.configurePassphraseCalls).toBe(0);
+      expect(passphraseEntry.authenticateCalls).toStrictEqual([]);
+      expect(passphraseEntry.configurePassphraseCalls).toBe(0);
+      const publicKey = await key.getPublicKey();
+      expect(passphraseEntry.authenticateCalls).toStrictEqual([
+        { keyId: "3c7bec5e-1f1d-4754-a1ce-3644ce1130f1", firstAttempt: true },
+      ]);
+      expect(passphraseEntry.configurePassphraseCalls).toBe(0);
 
-    expect(publicKey).toBeInstanceOf(PublicKey);
-    expect(publicKey.toHex("compressed")).toStrictEqual(
-      "03dfd25cea92d828d5c3b2836ffe5a843854285027b9de2dcca0920c0ff0beb5c4",
-    );
-  });
+      expect(publicKey).toBeInstanceOf(PublicKey);
+      expect(publicKey.toHex("compressed")).toStrictEqual(
+        "03dfd25cea92d828d5c3b2836ffe5a843854285027b9de2dcca0920c0ff0beb5c4",
+      );
+    });
 
-  test("sign()", async () => {
-    const pubKey = PublicKey.fromHex(
-      "03dfd25cea92d828d5c3b2836ffe5a843854285027b9de2dcca0920c0ff0beb5c4",
-      "compressed",
-    );
-    await fc.assert(
-      fc.asyncProperty(fc.uint8Array(), async (msg: Uint8Array) => {
-        const passphraseEntry = new MockPassphraseEntry();
-        const key = new Web3Account(pbkdf2KeyObject, passphraseEntry);
-        const sig = await key.sign(msg);
+    for (const [
+      kdf,
+      { keyObject, passphrase, privateKey, options },
+    ] of Object.entries(testVectors)) {
+      test(kdf, async () => {
+        const passphraseEntry = new MockPassphraseEntry(passphrase);
+        const key = new Web3Account(keyObject, passphraseEntry, options);
+
+        expect(passphraseEntry.authenticateCalls).toStrictEqual([]);
+        expect(passphraseEntry.configurePassphraseCalls).toBe(0);
+        const publicKey = await key.getPublicKey();
         expect(passphraseEntry.authenticateCalls).toStrictEqual([
-          { keyId: "3c7bec5e-1f1d-4754-a1ce-3644ce1130f1", firstAttempt: true },
+          { keyId: keyObject.id, firstAttempt: true },
         ]);
         expect(passphraseEntry.configurePassphraseCalls).toBe(0);
-        return await pubKey.verify(msg, sig);
-      }),
-    );
+
+        expect(publicKey).toBeInstanceOf(PublicKey);
+        expect(publicKey.toHex("compressed")).toStrictEqual(
+          await (await privateKey.getPublicKey()).toHex("compressed"),
+        );
+
+        if (options?.allowWeakPrivateKey) {
+          const keyWithStricterOptions = new Web3Account(
+            keyObject,
+            passphraseEntry,
+            {
+              allowWeakPrivateKey: false,
+            },
+          );
+          const publicKey2 = await key.getPublicKey();
+          expect(publicKey2).toBeInstanceOf(PublicKey);
+          expect(publicKey2.toHex("compressed")).toStrictEqual(
+            await (await privateKey.getPublicKey()).toHex("compressed"),
+          );
+        }
+      });
+    }
+  });
+
+  describe("sign()", () => {
+    test("fastcheck", async () => {
+      const pubKey = PublicKey.fromHex(
+        "03dfd25cea92d828d5c3b2836ffe5a843854285027b9de2dcca0920c0ff0beb5c4",
+        "compressed",
+      );
+      await fc.assert(
+        fc.asyncProperty(fc.uint8Array(), async (msg: Uint8Array) => {
+          const passphraseEntry = new MockPassphraseEntry();
+          const key = new Web3Account(pbkdf2KeyObject, passphraseEntry);
+          const sig = await key.sign(msg);
+          expect(passphraseEntry.authenticateCalls).toStrictEqual([
+            {
+              keyId: "3c7bec5e-1f1d-4754-a1ce-3644ce1130f1",
+              firstAttempt: true,
+            },
+          ]);
+          expect(passphraseEntry.configurePassphraseCalls).toBe(0);
+          return await pubKey.verify(msg, sig);
+        }),
+      );
+    });
+
+    for (const [
+      kdf,
+      { keyObject, passphrase, privateKey, options },
+    ] of Object.entries(testVectors)) {
+      const msg = new Uint8Array(128);
+      crypto.getRandomValues(msg);
+
+      test(kdf, async () => {
+        const passphraseEntry = new MockPassphraseEntry(passphrase);
+        const key = new Web3Account(keyObject, passphraseEntry, options);
+
+        expect(passphraseEntry.authenticateCalls).toStrictEqual([]);
+        expect(passphraseEntry.configurePassphraseCalls).toBe(0);
+        const sig = await key.sign(msg);
+        expect(passphraseEntry.authenticateCalls).toStrictEqual([
+          { keyId: keyObject.id, firstAttempt: true },
+        ]);
+        expect(passphraseEntry.configurePassphraseCalls).toBe(0);
+
+        const publicKey = await privateKey.getPublicKey();
+        expect(publicKey.verify(msg, sig)).toBeTruthy();
+
+        if (options?.allowWeakPrivateKey) {
+          const keyWithStricterOptions = new Web3Account(
+            keyObject,
+            passphraseEntry,
+            {
+              allowWeakPrivateKey: false,
+            },
+          );
+          expect(keyWithStricterOptions.sign(msg)).rejects.toThrowError(
+            WeakPrivateKeyError,
+          );
+        }
+      });
+    }
   });
 
   describe("exportPrivateKey()", () => {
-    for (const [kdf, { keyObject, passphrase, privateKey }] of Object.entries(
-      testVectors,
-    )) {
+    for (const [
+      kdf,
+      { keyObject, passphrase, privateKey, options },
+    ] of Object.entries(testVectors)) {
       test(kdf, async () => {
         const passphraseEntry = new MockPassphraseEntry(passphrase);
-        const key = new Web3Account(keyObject, passphraseEntry);
+        const key = new Web3Account(keyObject, passphraseEntry, options);
 
         expect(passphraseEntry.authenticateCalls).toStrictEqual([]);
         expect(passphraseEntry.configurePassphraseCalls).toBe(0);
@@ -130,6 +254,18 @@ describe("Web3Account", () => {
 
         expect(exported).toBeInstanceOf(RawPrivateKey);
         expect(exported.toBytes()).toStrictEqual(privateKey.toBytes());
+
+        if (options?.allowWeakPrivateKey) {
+          const keyWithStricterOptions = new Web3Account(
+            keyObject,
+            passphraseEntry,
+            {
+              allowWeakPrivateKey: false,
+            },
+          );
+          const exported2 = await keyWithStricterOptions.exportPrivateKey();
+          expect(exported2.toBytes()).toStrictEqual(privateKey.toBytes());
+        }
       });
     }
   });
@@ -156,13 +292,15 @@ test("encryptKeyObject()", async () => {
 });
 
 describe("decryptKeyObject()", async () => {
-  for (const [kdf, { keyObject, passphrase, privateKey }] of Object.entries(
-    testVectors,
-  )) {
+  for (const [
+    kdf,
+    { keyObject, passphrase, privateKey, options },
+  ] of Object.entries(testVectors)) {
     test(kdf, async () => {
       const { keyId, privateKey } = await decryptKeyObject(
         keyObject,
         passphrase,
+        options,
       );
       expect(keyId).toBe(keyObject.id);
       expect(privateKey.toBytes()).toStrictEqual(privateKey.toBytes());
@@ -170,6 +308,14 @@ describe("decryptKeyObject()", async () => {
       await expect(
         decryptKeyObject(keyObject, `wrong ${passphrase}`),
       ).rejects.toThrow(IncorrectPassphraseError);
+
+      if (options?.allowWeakPrivateKey) {
+        await expect(
+          decryptKeyObject(keyObject, passphrase, {
+            allowWeakPrivateKey: false,
+          }),
+        ).rejects.toThrow(WeakPrivateKeyError);
+      }
     });
   }
 });